% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/availableDates.R
\name{availableDates}
\alias{availableDates}
\title{TCGA datasets' releases dates}
\usage{
availableDates()
}
\value{
A vector of available dates to pass to the
\link{downloadTCGA} function.
}
\description{
Enables to check dates of TCGA datasets' releases.
}
\examples{
\dontrun{
availableDates()
}
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}};
\code{\link{availableDataSets}};
\code{\link{availableGenesNames}},
\code{\link{checkGenesNamesAvailability}};
\code{\link{calculate_cooccurence}};
\code{\link{checkDataSetsAvailability}};
\code{\link{downloadTCGA}}; \code{\link{infoTCGA}};
\code{\link{mergeTCGA}}; \code{\link{read.clinical}}
}

