% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/availableGenesNames.R
\name{checkGenesNamesAvailability}
\alias{availableGenesNames}
\alias{checkGenesNamesAvailability}
\title{TCGA genes' names and availability in \code{Merge_rnaseqv2__illuminahiseq_rnaseqv2__unc_edu__Level_3__RSEM_genes_normalized__data.Level_3} dataset}
\usage{
checkGenesNamesAvailability(rnaseqDir, genes)

availableGenesNames(rnaseqDir)
}
\arguments{
  \item{rnaseqDir}{TODO}

  \item{genes}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
\dontrun{

}
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}};
\code{\link{availableDataSets}};
\code{\link{availableDates}};
\code{\link{checkDataSetsAvailability}};
\code{\link{downloadTCGA}}; \code{\link{infoTCGA}},
\code{\link{mergaTCGA}},
\code{\link{mergeTCGA_clinical_rnaseq}};
\code{\link{mergeTCGA_clinical_mutations}};
\code{\link{read.clinical}}

Other RTCGA: \code{\link{RTCGA-package}};
\code{\link{availableDataSets}};
\code{\link{availableDates}};
\code{\link{checkDataSetsAvailability}};
\code{\link{downloadTCGA}}; \code{\link{infoTCGA}},
\code{\link{mergaTCGA}},
\code{\link{mergeTCGA_clinical_rnaseq}};
\code{\link{mergeTCGA_clinical_mutations}};
\code{\link{read.clinical}}
}

