% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkDataSetsAvailability.R
\name{checkDataSetsAvailability}
\alias{checkDataSetsAvailability}
\title{TCGA datasets' names availability}
\usage{
checkDataSetsAvailability(cancerTypes, pattern = "Merge_Clinical.Level_1",
  date = NULL)
}
\arguments{
  \item{cancerTypes}{A character vector containing
  abbreviation (Cohort code) of types of cancers to check
  for availability of datasets' name on
  \href{http://gdac.broadinstitute.org/}{http://gdac.broadinstitute.org/}.}

  \item{pattern}{A character vector of length 1 containing
  a part of a dataset's name to be checked for availability
  for current \code{date} parameter. By default phrase
  \code{"Merge_Clinical.Level_1"} is checked.}

  \item{date}{A \code{NULL} or character specifying from
  which date datasets' names should be checked for
  availability. By default (\code{date = NULL}) the newest
  available date is used. All available dates can be
  checked on
  \href{http://gdac.broadinstitute.org/runs/}{http://gdac.broadinstitute.org/runs/}
  or by using \link{availableDates} function. Required
  format \code{"YYYY-MM-DD"}.}
}
\value{
A vector of available datasets names to pass to the
\link{downloadTCGA} function.
}
\description{
Enables to check TCGA datasets' names availability for
current release date.
}
\examples{
\dontrun{
checkDataSetsAvailability( "BRCA" )
checkDataSetsAvailability( c("BRCA", "OV") )
checkDataSetsAvailability( "BRCA", "Mutation_Packager_Calls.Level" )
}
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}};
\code{\link{availableDataSets}};
\code{\link{availableDates}};
\code{\link{availableGenesNames}},
\code{\link{checkGenesNamesAvailability}};
\code{\link{downloadTCGA}}; \code{\link{infoTCGA}},
\code{\link{mergeTCGA}};
\code{\link{mergeTCGA_clinical_mutations}},
\code{\link{mergeTCGA_clinical_rnaseq}};
\code{\link{read.clinical}}
}

