% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/downloadTCGA.R
\name{downloadTCGA}
\alias{downloadTCGA}
\title{download}
\usage{
downloadTCGA(cancerTypes, additionalInfo = "Merge_Clinical.Level_1", destDir,
  date = NULL)
}
\description{
TODO
}
\examples{
\dontrun{

dir.create( "hre")

downloadTCGA( cancerTypes = "BRCA", additionalInfo = "miR_gene_expression",
destDir = "hre/" )

downloadTCGA( cancerTypes = c("BRCA", "OV"), destDir = "hre/" )
}
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}};
\code{\link{mergaTCGA}}
}

