% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installTCGA.R
\name{installTCGA}
\alias{installTCGA}
\title{Install packages from RTCGA family}
\usage{
installTCGA(packages = c("RTCGA.clinical", "RTCGA.mutations", "RTCGA.rnaseq",
  "RTCGA.RPPA", "RTCGA.mRNA", "RTCGA.CNV", "RTCGA.miRNASeq", "RTCGA.PANCAN12",
  "RTCGA.methylation"))
}
\arguments{
\item{packages}{A character specifing the names of the data packages to be installed. By default installs all packages.}
}
\description{
Function installs data packages from \href{https://github.com/RTCGA/}{https://github.com/RTCGA/}. Packages are listed \link{datasetsTCGA}.
}
\examples{
\dontrun{
installTCGA('RTCGA.clinical')

installTCGA()
}

}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}},
  \code{\link{checkTCGA}}, \code{\link{convertTCGA}},
  \code{\link{datasetsTCGA}}, \code{\link{downloadTCGA}},
  \code{\link{infoTCGA}}, \code{\link{readTCGA}}
}

