% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.clinical.R
\name{read.clinical}
\alias{read.clinical}
\title{Read from txt fo;e}
\usage{
read.clinical(clinicalDir, ...)
}
\arguments{
  \item{clinicalDir}{A directory to a
  \code{cancerType.clin.merged.txt} file. \code{cancerType}
  might be \code{BRCA, OV} etc.}
}
\value{
A data.frame with clinical data.
}
\description{
TODO
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}};
\code{\link{availableDataSets}};
\code{\link{availableDates}};
\code{\link{availableGenesNames}},
\code{\link{checkGenesNamesAvailability}};
\code{\link{calculate_cooccurence}};
\code{\link{checkDataSetsAvailability}};
\code{\link{downloadTCGA}}; \code{\link{infoTCGA}};
\code{\link{mergeTCGA}}
}

