% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTCGA-package.R
\docType{package}
\name{RTCGA-package}
\alias{RTCGA}
\alias{RTCGA-package}
\title{The Caner Genome Atlas data integration}
\description{
The Cancer Genome Atlas (TCGA) Data Portal provides a platform for researchers to search, download, and analyze data sets generated by TCGA. It contains clinical information, genomic characterization data, and high level sequence analysis of the tumor genomes. The key is to understand genomics to improve cancer care. RTCGA package offers download and integration of the variety and volume of TCGA data using patient barcode key, what enables easier data possession. This may have an benefcial infuence on impact on development of science and improvement of patients' treatment. Furthermore, RTCGA package transforms TCGA data to form which is convenient to use in R statistical package. Those data transformations can be a part of statistical analysis pipeline which can be more reproducible with RTCGA
}
\section{Issues}{


If you have any problems, issues or think that something is missing or is not
clear please post an issue on 
\href{https://github.com/RTCGA/RTCGA/issues}{https://github.com/RTCGA/RTCGA/issues}.
}
\examples{
\dontrun{
browseVignettes('RTCGA')
}
}
\author{
Marcin Kosinski [aut, cre] \email{ m.p.kosinski@gmail.com } \cr
Przemyslaw Biecek [aut] \email{ przemyslaw.biecek@gmail.com } \cr
Witold Chodor [ctb] \email{witoldchodor@gmail.com}
}
\seealso{
\pkg{RTCGA} website \href{http://rtcga.github.io/RTCGA}{http://rtcga.github.io/RTCGA}.

Other RTCGA: \code{\link{boxplotTCGA}},
  \code{\link{checkTCGA}}, \code{\link{convertTCGA}},
  \code{\link{createTCGA}}, \code{\link{datasetsTCGA}},
  \code{\link{downloadTCGA}},
  \code{\link{expressionsTCGA}}, \code{\link{heatmapTCGA}},
  \code{\link{infoTCGA}}, \code{\link{installTCGA}},
  \code{\link{kmTCGA}}, \code{\link{mutationsTCGA}},
  \code{\link{pcaTCGA}}, \code{\link{readTCGA}},
  \code{\link{survivalTCGA}}, \code{\link{theme_RTCGA}}
}

