% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalTCGA.R
\name{survivalTCGA}
\alias{survivalTCGA}
\title{Extract Survival Information From RTCGA.clinical Datasets}
\usage{
survivalTCGA(..., extract.cols = NULL, extract.names = FALSE)
}
\arguments{
\item{...}{A data.frame or data.frames from TCGA study containing clinical informations. See \link[RTCGA.clinical]{clinical}.}

\item{extract.cols}{A character specifing the names of extra columns to be extracted with survival information.}

\item{extract.names}{Logical, whether to extract names of passed data.frames in \code{...}.}
}
\value{
A data.frame containing information about times and censoring for specific \code{bcr_patient_barcode}.
}
\description{
Function restores codes and counts for each cohort from TCGA project.
}
\note{
Input data.frames should contain columns \code{patient.bcr_patient_barcode}, 
\code{patient.vital_status}, \code{patient.days_to_last_followup}, \code{patient.days_to_death}. 
It is recommended to use datasets from \link[RTCGA.clinical]{clinical}.
}
\examples{

## Extracting Survival Data
library(RTCGA.clinical)
survivalTCGA(BRCA.clinical, OV.clinical, extract.cols = "admin.disease_code") -> BRCAOV.survInfo
library(dplyr)
survivalTCGA(HNSC.clinical, 
   extract.cols = c("patient.drugs.drug.therapy_types.therapy_type",
   								 "admin.disease_code")) \%>\%
   filter(patient.drugs.drug.therapy_types.therapy_type ==
          "chemotherapy")-> HNSC.survInfo
          
## Kaplan-Meier Survival Curves
library(survminer)
library(survival)
ggsurvplot(survfit(Surv(times, patient.vital_status)~admin.disease_code,
									 data = BRCAOV.survInfo),
				 risk.table = TRUE,
					 ggtheme = theme_RTCGA(base_size = 16,
					 											 base_family = "serif"),
					 break.time.by = 800,
					 palette = c("#FF9E29", "#86AA00"))

}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}},
  \code{\link{checkTCGA}}, \code{\link{convertTCGA}},
  \code{\link{datasetsTCGA}}, \code{\link{downloadTCGA}},
  \code{\link{expressionTCGA}}, \code{\link{infoTCGA}},
  \code{\link{installTCGA}}, \code{\link{readTCGA}},
  \code{\link{theme_RTCGA}}
}

