% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_RTCGA.R
\name{theme_RTCGA}
\alias{theme_RTCGA}
\title{RTCGA Theme For ggplot2}
\usage{
theme_RTCGA(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Additional theme for \link[ggplot2]{ggtheme}.
}
\examples{


library(RTCGA.clinical)
survivalTCGA(BRCA.clinical, OV.clinical, extract.names = TRUE) -> BRCAOV.survInfo
library(survminer)
library(survival)
ggsurvplot(survfit(Surv(times, patient.vital_status)~dataset,
									 data = BRCAOV.survInfo),
				 risk.table = TRUE,
					 ggtheme = theme_RTCGA(base_size = 16,
					 											 base_family = "serif"),
					 break.time.by = 800,
					 palette = c("#FF9E29", "#86AA00"))
					 
}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}},
  \code{\link{checkTCGA}}, \code{\link{convertTCGA}},
  \code{\link{datasetsTCGA}}, \code{\link{downloadTCGA}},
  \code{\link{expressionTCGA}}, \code{\link{infoTCGA}},
  \code{\link{installTCGA}}, \code{\link{readTCGA}},
  \code{\link{survivalTCGA}}
}

