% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/availableDataSets.R
\name{availableDataSets}
\alias{availableDataSets}
\title{TCGA datasets' names}
\usage{
availableDataSets(cancerType, date = NULL)
}
\arguments{
  \item{cancerType}{A character of length 1 containing
  abbreviation (Cohort code) of types of cancers to check
  for available datasets' names on
  \href{http://gdac.broadinstitute.org/}{http://gdac.broadinstitute.org/}.}

  \item{date}{A \code{NULL} or character specifying from
  which date datasets' names should be checked. By default
  (\code{date = NULL}) the newest available date is used.
  All available dates can be checked on
  \href{http://gdac.broadinstitute.org/runs/}{http://gdac.broadinstitute.org/runs/}
  or by using \link{availableDates} function. Required
  format \code{"YYYY-MM-DD"}.}
}
\value{
A vector of available datasets names to pass to the
\link{downloadTCGA} function.
}
\description{
Enables to check TCGA datasets' names for current release
date.
}
\examples{
\dontrun{
availableDataSets( "BRCA" )
availableDataSets( "OV", availableDates()[5] ) # error
}
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}};
\code{\link{availableDates}};
\code{\link{availableGenesNames}},
\code{\link{checkGenesNamesAvailability}};
\code{\link{checkDataSetsAvailability}};
\code{\link{downloadTCGA}}; \code{\link{mergaTCGA}},
\code{\link{mergeTCGA_clinical_mutations}},
\code{\link{mergeTCGA_clinical_rnaseq}},
\code{\link{prepareTCGA_mutations_for_merging}}
}

