% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkTCGA.R
\name{checkTCGA}
\alias{checkTCGA}
\title{Information about datasets from TCGA project}
\usage{
checkTCGA(what, cancerType, date = NULL)
}
\arguments{
\item{what}{One of \code{DataSets} or \code{Dates}.}

\item{cancerType}{A character of length 1 containing abbreviation (Cohort code - \href{http://gdac.broadinstitute.org/}{http://gdac.broadinstitute.org/})
of types of cancers to check for.}

\item{date}{A \code{NULL} or character specifying from which date informations should be checked.
By default (\code{date = NULL}) the newest available date is used. All available dates can be checked on
\href{http://gdac.broadinstitute.org/runs/}{http://gdac.broadinstitute.org/runs/} or by using \code{checkTCGA('Dates')}
function. Required format \code{'YYYY-MM-DD'}.}
}
\value{
\itemize{
     \item If \code{what='DataSets'} a vector of available datasets' names to pass to the \link{downloadTCGA} function.
     \item If \code{what='Dates'} a vector of available dates to pass to the \link{downloadTCGA} function.
     }
}
\description{
The \code{checkTCGA} function let's to check

\itemize{
     \item \code{DataSets}: TCGA datasets' names for current release date and cohort.
     \item \code{Dates}: TCGA datasets' dates of release.
 }
}
\details{
\itemize{
     \item If \code{what='DataSets'} enables to check TCGA datasets' names for current release date and cohort.
     \item If \code{what='Dates'} enables to check dates of TCGA datasets' releases.
     }
}
\examples{
\dontrun{

#############################

# names for current release date and cohort
checkTCGA('DataSets', 'BRCA' )
checkTCGA('DataSets', 'OV', tail(checkTCGA('Dates'))[1] )
#checkTCGA('DataSets', 'OV', checkTCGA('Dates')[5] ) # error

# dates of TCGA datasets' releases.
checkTCGA('Dates')

#############################

# TCGA datasets' names availability for
# current release date and cancer type.

releaseDate <- '2015-06-01'
cancerTypes <- c('OV', 'BRCA')

cancerTypes \%>\% sapply(function(element){
  grep(x = checkTCGA('DataSets', element, releaseDate),
      pattern = 'humanmethylation450', value = TRUE) \%>\%
       as.vector()
       })

#############################

# TCGA genes' names and availability
# in Merge_rnaseqv2__... dataset
dir.create('data2')
sapply( cancerTypes, function(element){
tryCatch({
    downloadTCGA( cancerTypes = element,
                  dataSet = paste0('rnaseqv2__illuminahiseq_rnaseqv2__unc',
                  '_edu__Level_3__RSEM_genes_normalized__data.Level'),
                  destDir = 'data2',
                  date = releaseDate )},
    error = function(cond){
        cat('Error: Maybe there weren't rnaseq data for ', element, ' cancer.\\n')
    }
)
})

# Paths to rna-seq data

sapply( cohorts, function( element ){
folder <- grep( paste0( '(_',element,'\\\\.', '|','_',element,'-FFPE)', '.*rnaseqv2'),
               list.files('data2/'), value = TRUE)
file <- grep( paste0(element, '.*rnaseqv2'), list.files( paste0( 'data2/',folder ) ),
              value = TRUE)
path <- paste0( 'data2/', folder, '/', file )
assign( value = path, x = paste0(element, '.rnaseq.path'), envir = .GlobalEnv)
})

rnaseqDir <- 'OV.rnaseq'


fread(rnaseqDir, select = c(1),
      data.table = FALSE,
      colClasses = 'character')[-1, 1]

}
}
\seealso{
Other RTCGA: \code{\link{RTCGA}},
  \code{\link{RTCGA-package}}; \code{\link{datasetsTCGA}};
  \code{\link{downloadTCGA}}; \code{\link{infoTCGA}};
  \code{\link{readTCGA}}
}

