% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasetsTCGA.R
\name{datasetsTCGA}
\alias{datasetsTCGA}
\title{RTCGA.data - The Family of R Packages with Data from The Cancer Genome Atlas Study}
\description{
Snapshots of the clinical, mutations, cnvs and rnaseq datasets from the \code{2015-06-01}
release date are included in the \code{RTCGA.data} family that contains 4 packages:
\itemize{
 \item \pkg{RTCGA.rnaseq}
 \item \pkg{RTCGA.clinical}
 \item \pkg{RTCGA.mutations}
 \item \pkg{RTCGA.cnv}
 }
}
\details{
For more detailed information visit \pkg{RTCGA.data}
\href{https://mi2-warsaw.github.io/RTCGA.data}{website}.
}
\examples{
# installation of packages containing snapshots
# of TCGA project's datasets

\dontrun{
source('http://bioconductor.org/biocLite.R')
biocLite(RTCGA.clinical)
biocLite(RTCGA.mutations)
biocLite(RTCGA.rnaseq)
biocLite(RTCGA.cnv)

# use cases and examples + more data info
browseVignettes('RTCGA')
}
}
\author{
Marcin Kosinski [aut, cre] \email{ m.p.kosinski@gmail.com } \cr
Przemyslaw Biecek [aut] \email{ przemyslaw.biecek@gmail.com }
}
\seealso{
Other RTCGA: \code{\link{RTCGA}},
  \code{\link{RTCGA-package}}; \code{\link{checkTCGA}};
  \code{\link{downloadTCGA}}; \code{\link{infoTCGA}};
  \code{\link{readTCGA}}
}

