% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/downloadTCGA.R
\name{downloadTCGA}
\alias{downloadTCGA}
\title{download}
\usage{
downloadTCGA(cancerTypes, additionalInfo = "Merge_Clinical.Level_1",
  writeDirectory)
}
\description{
TODO
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}};
\code{\link{mergaTCGA}}
}

