% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downloadTCGA.R
\name{downloadTCGA}
\alias{downloadTCGA}
\title{Download TCGA data}
\usage{
downloadTCGA(cancerTypes, dataSet = "Merge_Clinical.Level_1", destDir,
  date = NULL, untarFile = TRUE, removeTar = TRUE)
}
\arguments{
\item{cancerTypes}{A character vector containing abbreviations (Cohort code) of types of cancers to download from
\href{http://gdac.broadinstitute.org/}{http://gdac.broadinstitute.org/}. For easy access from R check details below.}

\item{dataSet}{A part of the name of dataSet to be downloaded from \href{http://gdac.broadinstitute.org/runs/}{http://gdac.broadinstitute.org/runs/}. By default the Merged Clinical dataSet is downloaded (value \code{dataSet = 'Merge_Clinical.Level_1'}). Available datasets' names can be checked using \link{checkTCGA} function.}

\item{destDir}{A character specifying a directory into which \code{dataSet}s will be downloaded.}

\item{date}{A \code{NULL} or character specifying from which date \code{dataSet}s should be downloaded.
By default (\code{date = NULL}) the newest available date is used. All available dates can be checked on
\href{http://gdac.broadinstitute.org/runs/}{http://gdac.broadinstitute.org/runs/} or by using \link{checkTCGA}
function. Required format \code{'YYYY-MM-DD'}.}

\item{untarFile}{Logical - should the downloaded file be untarred. Default is \code{TRUE}.}

\item{removeTar}{Logical - should the downloaded \code{.tar} file be removed after untarring.
Default is \code{TRUE}.}
}
\value{
No values. It only downloads files.
}
\description{
Enables to download TCGA data from specified dates of releases of concrete Cohorts of cancer types.
Pass a name of required dataset to the \code{dataSet} parameter. By default the Merged Clinical
dataSet is downloaded (value \code{dataSet = 'Merge_Clinical.Level_1'}) from the newest available date of the release.
}
\details{
All cohort names can be checked using: \code{ sub( x = names( infoTCGA() ), '-counts', '' )}.
}
\examples{
\dontrun{
dir.create( 'hre')

downloadTCGA( cancerTypes = 'ACC', dataSet = 'miR_gene_expression',
destDir = 'hre', date =  tail( checkTCGA('Dates'), 2 )[1] )


downloadTCGA( cancerTypes = c('BRCA', 'OV'), destDir = 'hre',
 date = tail( checkTCGA('Dates'), 2 )[1] )
}
}
\seealso{
Other RTCGA: \code{\link{RTCGA}},
  \code{\link{RTCGA-package}}; \code{\link{checkTCGA}};
  \code{\link{datasetsTCGA}}; \code{\link{infoTCGA}};
  \code{\link{readTCGA}}
}

