% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/infoTCGA.R
\name{infoTCGA}
\alias{infoTCGA}
\title{Information about cohorts from TCGA project}
\usage{
infoTCGA()
}
\value{
A list with a tabular information from
\href{http://gdac.broadinstitute.org/}{http://gdac.broadinstitute.org/}.
}
\description{
Function restores codes and counts for each cohort from
TCGA project.
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}};
\code{\link{availableDataSets}};
\code{\link{availableDates}};
\code{\link{availableGenesNames}},
\code{\link{checkGenesNamesAvailability}};
\code{\link{checkDataSetsAvailability}};
\code{\link{downloadTCGA}}; \code{\link{mergeTCGA}};
\code{\link{read.clinical}}
}

