% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoTCGA.R
\name{infoTCGA}
\alias{infoTCGA}
\title{Information about cohorts from TCGA project}
\usage{
infoTCGA()
}
\value{
A list with a tabular information from \href{http://gdac.broadinstitute.org/}{http://gdac.broadinstitute.org/}.
}
\description{
Function restores codes and counts for each cohort from TCGA project.
}
\examples{

infoTCGA()
library(magrittr)
(cohorts <- infoTCGA() \%>\% 
rownames() \%>\% 
   sub('-counts', '', x=.))

}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}},
  \code{\link{boxplotTCGA}}, \code{\link{checkTCGA}},
  \code{\link{convertTCGA}}, \code{\link{datasetsTCGA}},
  \code{\link{downloadTCGA}},
  \code{\link{expressionsTCGA}}, \code{\link{installTCGA}},
  \code{\link{kmTCGA}}, \code{\link{mutationsTCGA}},
  \code{\link{pcaTCGA}}, \code{\link{readTCGA}},
  \code{\link{survivalTCGA}}, \code{\link{theme_RTCGA}}
}

