% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infoTCGA.R
\name{infoTCGA}
\alias{infoTCGA}
\title{Information about cohorts from TCGA project}
\usage{
infoTCGA()
}
\value{
A list with a tabular information from \href{http://gdac.broadinstitute.org/}{http://gdac.broadinstitute.org/}.
}
\description{
Function restores codes and counts for each cohort from TCGA project.
}
\examples{
infoTCGA()

(cohorts <- infoTCGA() \%>\%
rownames() \%>\%
   sub('-counts', '', x=.))
}
\seealso{
Other RTCGA: \code{\link{RTCGA}},
  \code{\link{RTCGA-package}}; \code{\link{checkTCGA}};
  \code{\link{datasetsTCGA}}; \code{\link{downloadTCGA}};
  \code{\link{readTCGA}}
}

