% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mergeTCGA.R
\name{mergaTCGA}
\alias{mergaTCGA}
\title{Merge}
\usage{
mergaTCGA(clinical, names)
}
\description{
TODO
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}};
\code{\link{availableDataSets}};
\code{\link{availableDates}};
\code{\link{checkDataSetsAvailability}};
\code{\link{checkGenesNamesAvailability}};
\code{\link{downloadTCGA}}
}

