% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readTCGA.R
\name{readTCGA}
\alias{readTCGA}
\title{Read TCGA data to the tidy format}
\usage{
readTCGA(path, dataType, ...)
}
\arguments{
\item{path}{If \code{dataType = "clinical"} a directory to a \code{cancerType.clin.merged.txt} file.
\code{cancerType} might be \code{BRCA, OV} etc.}

\item{dataType}{TODO}

\item{...}{Further arguments passed to the \link{as.data.frame}.}
}
\value{
An output:
\itemize{
     \item If \code{dataType = "clinical"} a \code{data.frame} with clinical data.
}
}
\description{
TODO
}
\examples{
\dontrun{

##############
##### clinical
##############

# downloading clinical data
downloadTCGA( cancerTypes = c("BRCA", "OV"), destDir = "hre/" )

# untarring files
list.files( "hre/") \%>\%
paste0( "hre/", .) \%>\%
   sapply( untar, exdir = "hre/" )

# removing *.tar.gz files
list.files( "hre/") \%>\%
paste0( "hre/", .) \%>\%
    grep( pattern = "tar.gz", x = ., value = TRUE) \%>\%
    sapply( file.remove )

# reading datasets
sapply( c("BRCA", "OV"), function( element ){
folder <- grep( paste0( "(_",element,"\\\\.", "|","_",element,"-FFPE)", ".*Clinical"),
      list.files("hre/"),value = TRUE)
path <- paste0( "hre/", folder, "/", element, ".clin.merged.txt")
assign( value = readTCGA( path, "clinical" ),
         x = paste0(element, ".clin.data"), envir = .GlobalEnv)
         })
}
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}};
  \code{\link{checkTCGA}}; \code{\link{downloadTCGA}};
  \code{\link{infoTCGA}}
}

