% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTCGA.R
\name{readTCGA}
\alias{readTCGA}
\title{Read TCGA data to the tidy format}
\usage{
readTCGA(path, dataType, ...)
}
\arguments{
\item{path}{See details and examples.}

\item{dataType}{One of \code{'clinical', 'rnaseq', 'mutations', 'RPPA', 'mRNA', 'miRNASeq', 'methylation'} depending on which type of data user is trying to read in the tidy format.}

\item{...}{Further arguments passed to the \link{as.data.frame}.}
}
\value{
An output:
\itemize{
     \item If \code{dataType = 'clinical'} a \code{data.frame} with clinical data.
     \item If \code{dataType = 'rnaseq'} a \code{data.frame} with rnaseq data.
     \item If \code{dataType = 'mutations'} a \code{data.frame} with mutations data.
     \item If \code{dataType = 'RPPA'} a \code{data.frame} with RPPA data.
     \item If \code{dataType = 'mRNA'} a \code{data.frame} with mRNA data.
     \item If \code{dataType = 'miRNASeq'} a \code{data.frame} with miRNASeq data.
     \item If \code{dataType = 'methylation'} a \code{data.frame} with methylation data.
}
}
\description{
\code{readTCGA} function allows to read unzipped files: 
\itemize{
     \item clinical data - \code{Merge_Clinical.Level_1} 
     \item rnaseq data (genes' expressions) - \code{rnaseqv2__illuminahiseq_rnaseqv2}
     \item genes' mutations data - \code{Mutation_Packager_Calls.Level}
     \item Reverse phase protein array data (RPPA) - \code{protein_normalization__data.Level_3}
     \item Merge transcriptome agilent data (mRNA) - \code{Merge_transcriptome__agilentg4502a_07_3__unc_edu__Level_3__unc_lowess_normalization_gene_level__data.Level_3}
     \item miRNASeq data - \code{Merge_mirnaseq__illuminaga_mirnaseq__bcgsc_ca__Level_3__miR_gene_expression__data.Level_3} or \code{"Merge_mirnaseq__illuminahiseq_mirnaseq__bcgsc_ca__Level_3__miR_gene_expression__data.Level_3"}
     \item methylation data - \code{Merge_methylation__humanmethylation27} 
     }
from TCGA project. Those files can be easily downloded with \link{downloadTCGA} function. See examples.
}
\details{
All cohort names can be checked using: \code{ sub( x = names( infoTCGA() ), '-counts', '')}.

Parameter \code{path} specification: 
\itemize{
\item If \code{dataType = 'clinical'} a path to a \code{cancerType.clin.merged.txt} file. 
\item If \code{dataType = 'mutations'} a path to the unzziped folder \code{Mutation_Packager_Calls.Level} containing \code{.maf} files.
\item If \code{dataType = 'rnaseq'} a path to the uzziped file \code{rnaseqv2__illuminahiseq_rnaseqv2__unc_edu__Level_3__RSEM_genes_normalized__data.Level}.
\item If \code{dataType = 'RPPA'} a path to the unzipped file in folder \code{protein_normalization__data.Level_3}.
\item If \code{dataType = 'mRNA'} a path to the unzipped file \code{cancerType.transcriptome__agilentg4502a_07_3__unc_edu__Level_3__unc_lowess_normalization_gene_level__data.data.txt}
\item If \code{dataType = 'miRNASeq'} a path to unzipped files \code{cancerType.mirnaseq__illuminahiseq_mirnaseq__bcgsc_ca__Level_3__miR_gene_expression__data.data.txt} or \code{cancerType.mirnaseq__illuminaga_mirnaseq__bcgsc_ca__Level_3__miR_gene_expression__data.data.txt}
\item If \code{dataType = 'methylation'} a path to unzipped files \code{cancerType.methylation__humanmethylation27__jhu_usc_edu__Level_3__within_bioassay_data_set_function__data.data.txt}
}
}
\examples{

\dontrun{ 

##############
##### clinical
##############

dir.create('hre')

# downloading clinical data
downloadTCGA( cancerTypes = c('BRCA', 'OV'), destDir = 'hre' )

    
# reading datasets    
sapply( c('BRCA', 'OV'), function( element ){
folder <- grep( paste0( '(_',element,'\\\\.', '|','_',element,'-FFPE)', '.*Clinical'),
      list.files('hre/'),value = TRUE)
path <- paste0( 'hre/', folder, '/', element, '.clin.merged.txt')
assign( value = readTCGA( path, 'clinical' ), 
         x = paste0(element, '.clin.data'), envir = .GlobalEnv)
         })
     
##############
##### rnaseq
##############

dir.create('data2')

# downloading rnaseq data
downloadTCGA( cancerTypes = 'BRCA', 
dataSet = 'rnaseqv2__illuminahiseq_rnaseqv2__unc_edu__Level_3__RSEM_genes_normalized__data.Level',
destDir = 'data2' )

# shortening paths and directories
list.files( 'data2/') \%>\% 
    file.path( 'data2', .) \%>\%
    file.rename( to = substr(.,start=1,stop=50))

# reading data
list.files( 'data2/') \%>\% 
    file.path( 'data2', .) -> folder

folder \%>\%
    list.files \%>\%
    file.path( folder, .) \%>\%
    grep( pattern = 'illuminahiseq', x = ., value = TRUE) -> pathRNA
readTCGA( path = pathRNA, dataType = 'rnaseq' ) -> my_data


##############
##### mutations
##############

dir.create('data3')


downloadTCGA( cancerTypes = 'OV', 
              dataSet = 'Mutation_Packager_Calls.Level',
              destDir = 'data3' )

# reading data
list.files( 'data3/') \%>\% 
    file.path( 'data3', .) -> folder

readTCGA(folder, 'mutations') -> mut_file

}

}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}

Witold Chodor, \email{witoldchodor@gmail.com}
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}},
  \code{\link{checkTCGA}}, \code{\link{convertTCGA}},
  \code{\link{datasetsTCGA}}, \code{\link{downloadTCGA}},
  \code{\link{infoTCGA}}, \code{\link{installTCGA}}
}

