% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RTCGA-package.R
\docType{package}
\name{RTCGA-package}
\alias{RTCGA-package}
\title{The Caner Genome Atlas data integration}
\description{
The Cancer Genome Atlas (TCGA) Data Portal provides a
platform for researchers to search, download, and analyze
data sets generated by TCGA. It contains clinical
information, genomic characterization data, and high level
sequence analysis of the tumor genomes[1]. The key is to
understand genomics to improve cancer care. RTCGA package
offers download and integration of the variety and volume
of TCGA data using patient barcode key, what enables easier
data possession. This may have an benefcial infuence on
impact on development of science and improvement of
patients' treatment. RTCGA is an open-source R package,
available to download from Bioconductor[2]. Furthermore,
RTCGA package transforms TCGA data to form which is
convenient to use in R statistical package. Those data
transformations can be a part of statistical analysis
pipeline which can be more reproducible with RTCGA
}
\details{
For more detailed information visit \pkg{RTCGA} \code{wiki}
on
\href{https://github.com/MarcinKosinski/RTCGA/wiki}{Github}.
}
\author{
TODO [aut, cre] \email{ 1 } \cr TODO [aut] \email{ 2 } TODO
[aut] \email{ 3 }
}
\seealso{
Other RTCGA: \code{\link{availableDataSets}};
\code{\link{availableDates}};
\code{\link{availableGenesNames}},
\code{\link{checkGenesNamesAvailability}};
\code{\link{checkDataSetsAvailability}};
\code{\link{downloadTCGA}}; \code{\link{mergaTCGA}},
\code{\link{mergeTCGA_clinical_rnaseq}}
}

