% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalTCGA.R
\name{survivalTCGA}
\alias{survivalTCGA}
\title{Extract Survival Information From RTCGA.clinical Datasets}
\usage{
survivalTCGA(..., extra.cols = NULL)
}
\arguments{
\item{...}{A data.frame or data.frames containing from TCGA study containing clinical informations. See \link[RTCGA.clinical]{clinical}.}

\item{extra.cols}{A character specifing the names of extra columns to be extracted with survival information.}
}
\value{
A data.frame containing information about times and censoring for specific \code{bcr_patient_barcode}.
}
\description{
Function restores codes and counts for each cohort from TCGA project.
}
\note{
Input data.frames should contain columns \code{patient.bcr_patient_barcode}, 
\code{patient.vital_status}, \code{patient.days_to_last_followup}, \code{patient.days_to_death}. 
It is recommended to use datasets from \link[RTCGA.clinical]{clinical}.
}
\examples{

library(RTCGA.clinical)
survivalTCGA(BRCA.clinical, OV.clinical) -> BRCAOV.survInfo
library(dplyr)
survivalTCGA(HNSC.clinical, 
   extra.cols = "patient.drugs.drug.therapy_types.therapy_type") \%>\%
   filter(patient.drugs.drug.therapy_types.therapy_type ==
          "chemotherapy")-> HNSC.survInfo

}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}},
  \code{\link{checkTCGA}}, \code{\link{convertTCGA}},
  \code{\link{datasetsTCGA}}, \code{\link{downloadTCGA}},
  \code{\link{infoTCGA}}, \code{\link{installTCGA}},
  \code{\link{readTCGA}}
}

