% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalTCGA.R
\name{survivalTCGA}
\alias{survivalTCGA}
\title{Extract Survival Information From RTCGA.clinical Datasets}
\usage{
survivalTCGA(..., extract.cols = NULL, extract.names = FALSE)
}
\arguments{
\item{...}{A data.frame or data.frames from TCGA study containing clinical informations. See \link[RTCGA.clinical]{clinical}. 
Or for plot method a further arguments passed to \link[survminer]{ggsurvival}.}

\item{extract.cols}{A character specifing the names of extra columns to be extracted with survival information.}

\item{extract.names}{Logical, whether to extract names of passed data.frames in \code{...}.}

\item{x}{Object of class \code{survivalTCGA}.}

\item{times}{The name of time variable.}

\item{status}{The name of status variable.}

\item{explanatory.names}{Names of explanatory variables to use in survival curves plot.}

\item{title}{Title of the plot.}

\item{risk.table}{Whether to show risk tables.}

\item{conf.int}{Whether to show confidence intervals.}
}
\value{
A data.frame containing information about times and censoring for specific \code{bcr_patient_barcode}.
}
\description{
Function restores codes and counts for each cohort from TCGA project.
}
\note{
Input data.frames should contain columns \code{patient.bcr_patient_barcode}, 
\code{patient.vital_status}, \code{patient.days_to_last_followup}, \code{patient.days_to_death}. 
It is recommended to use datasets from \link[RTCGA.clinical]{clinical}.
}
\examples{

## Extracting Survival Data
library(RTCGA.clinical)
survivalTCGA(BRCA.clinical, OV.clinical, extract.cols = "admin.disease_code") -> BRCAOV.survInfo
library(dplyr)

BRCA.clinical \%>\%
    filter(patient.drugs.drug.therapy_types.therapy_type \%in\%
               c("chemotherapy", "hormone therapy")) \%>\%
    rename(therapy = patient.drugs.drug.therapy_types.therapy_type) \%>\%
    survivalTCGA(BRCA.clinical, 
                 extract.cols = c("therapy"))  -> BRCA.survInfo.chemo

## Kaplan-Meier Survival Curves
plot(BRCAOV.survInfo)


plot(BRCA.survInfo.chemo)

}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}},
  \code{\link{checkTCGA}}, \code{\link{convertTCGA}},
  \code{\link{datasetsTCGA}}, \code{\link{downloadTCGA}},
  \code{\link{expressionTCGA}}, \code{\link{infoTCGA}},
  \code{\link{installTCGA}}, \code{\link{readTCGA}},
  \code{\link{theme_RTCGA}}
}

