% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalTCGA.R
\name{survivalTCGA}
\alias{survivalTCGA}
\title{Extract Survival Information From RTCGA.clinical Datasets}
\usage{
survivalTCGA(..., extract.cols = NULL, extract.names = FALSE)
}
\arguments{
\item{...}{A data.frame or data.frames from TCGA study containing clinical informations. See \link[RTCGA.clinical]{clinical}.}

\item{extract.cols}{A character specifing the names of extra columns to be extracted with survival information.}

\item{extract.names}{Logical, whether to extract names of passed data.frames in \code{...}.}
}
\value{
A data.frame containing information about times and censoring for specific \code{bcr_patient_barcode}.
}
\description{
Extracts survival information from clicnial datasets from TCGA project.
}
\note{
Input data.frames should contain columns \code{patient.bcr_patient_barcode}, 
\code{patient.vital_status}, \code{patient.days_to_last_followup}, \code{patient.days_to_death}. 
It is recommended to use datasets from \link[RTCGA.clinical]{clinical}.
}
\examples{

## Extracting Survival Data
library(RTCGA.clinical)
survivalTCGA(BRCA.clinical, OV.clinical, extract.cols = "admin.disease_code") -> BRCAOV.survInfo

# first munge data, then extract survival info
library(dplyr)
BRCA.clinical \%>\%
    filter(patient.drugs.drug.therapy_types.therapy_type \%in\%
               c("chemotherapy", "hormone therapy")) \%>\%
    rename(therapy = patient.drugs.drug.therapy_types.therapy_type) \%>\%
    survivalTCGA(extract.cols = c("therapy"))  -> BRCA.survInfo.chemo
                 
# first extract survival info, then munge data                  
    survivalTCGA(BRCA.clinical, 
                 extract.cols = c("patient.drugs.drug.therapy_types.therapy_type"))  \%>\%
    filter(patient.drugs.drug.therapy_types.therapy_type \%in\%
               c("chemotherapy", "hormone therapy")) \%>\%
    rename(therapy = patient.drugs.drug.therapy_types.therapy_type) -> BRCA.survInfo.chemo

## Kaplan-Meier Survival Curves
kmTCGA(BRCAOV.survInfo, explanatory.names = "admin.disease_code",  pval = TRUE)

kmTCGA(BRCAOV.survInfo, explanatory.names = "admin.disease_code", main = "",
       xlim = c(0,4000))
 
kmTCGA(BRCA.survInfo.chemo, explanatory.names = "therapy", xlim = c(0, 3000), conf.int = FALSE)

}
\author{
Marcin Kosinski, \email{m.p.kosinski@gmail.com}
}
\seealso{
Other RTCGA: \code{\link{RTCGA-package}},
  \code{\link{checkTCGA}}, \code{\link{convertTCGA}},
  \code{\link{datasetsTCGA}}, \code{\link{downloadTCGA}},
  \code{\link{expressionsTCGA}}, \code{\link{infoTCGA}},
  \code{\link{installTCGA}}, \code{\link{kmTCGA}},
  \code{\link{pcaTCGA}}, \code{\link{readTCGA}},
  \code{\link{theme_RTCGA}}
}

